+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [1.20.2] - October 9, 2024
 - Updated how Keyless YAML (`_`) works - it now creates a root key of `_` regardless of contents
+- Added support for a `filters` lookup to dynamically use filters
 - Jinja2 Extension `jinja2.ext.loopcontrols` is now enabled by default
 
 ### [1.20.1] - May 15, 2024